+++ src/app/core/models/credential-type-map.ts
@@ -5,9 +5,9 @@ export interface CredentialMapConfig {
 }
 
 export const CredentialTypeMap: Record<CredentialType, CredentialMapConfig> = {
-  'learcredential.employee.w3c.1': { icon: 'assets/icons/LearCredentialEmployee.png' },
+  'learcredential.employee.w3c.4': { icon: 'assets/icons/LearCredentialEmployee.png' },
   'learcredential.employee.sd.1':  { icon: 'assets/icons/LearCredentialEmployee.png' },
-  'learcredential.machine.w3c.1':  { icon: 'assets/icons/LearCredentialMachine.png' },
+  'learcredential.machine.w3c.3':  { icon: 'assets/icons/LearCredentialMachine.png' },
   'learcredential.machine.sd.1':   { icon: 'assets/icons/LearCredentialMachine.png' },
   'gx.labelcredential.w3c.1':      { icon: 'assets/icons/LabelCredential.png' },
   'doctorid.sd.1':                  { icon: 'assets/icons/LearCredentialEmployee.png' },