+++ src/app/core/models/verifiable-credential.ts
@@ -18,9 +18,9 @@ export interface VerifiableCredential {
 
 // Must match credential_configuration_id in assets/schemas/*.json
 export const CREDENTIAL_TYPES_ARRAY = [
-  'learcredential.employee.w3c.1',
+  'learcredential.employee.w3c.4',
   'learcredential.employee.sd.1',
-  'learcredential.machine.w3c.1',
+  'learcredential.machine.w3c.3',
   'learcredential.machine.sd.1',
   'gx.labelcredential.w3c.1',
   'doctorid.sd.1',